% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.character.slidepzl_state}
\alias{as.character.slidepzl_state}
\title{convert a state into a vector of character string}
\usage{
\method{as.character}{slidepzl_state}(state)
}
\arguments{
\item{state}{an object of 'slidepzl_state' class.}
}
\value{
a character vector
}
\description{
Internal. Convert a state into a vector of character string.
It is concatenated and used for a search index of the state.
}
