% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.matrix.slidepzl_state}
\alias{as.matrix.slidepzl_state}
\title{convert a state into a character matrix.}
\usage{
\method{as.matrix}{slidepzl_state}(state, setting, check_valid = T)
}
\arguments{
\item{state}{a object of 'slidepzl_state' class.}

\item{setting}{a object of 'slidepzl_setting' class.}

\item{check_valid}{logical. check the validity of the state?}
}
\value{
a character matrix.
}
\description{
convert a state into a character matrix.
}
