% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graph.R
\name{getAllShortestPaths}
\alias{getAllShortestPaths}
\title{get all of shortest paths to target}
\usage{
getAllShortestPaths(g, target = NULL)
}
\arguments{
\item{g}{an 'igraph' object generated by \code{makeGraph} function.}

\item{target}{a numerical vector, which gives
IDs of target states. NULL means IDs of goal states.}
}
\value{
a list with following elements:
        \describe{
          \item{state}{a list of character vectors.
          ith element gives states in the ith path.}
          \item{transition}{a list of character vectors.
          ith element gives transitions in the ith path.}
        }
}
\description{
Get all of shortest paths from initial states to goal states
(or to specified states). It may take a long time.
}
\details{
see \code{\link{makeGraph}} for example code.
}
