% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_state.R
\name{getNumStates}
\alias{getNumStates}
\title{return number of possible states}
\usage{
getNumStates(state, setting, verbose = T)
}
\arguments{
\item{state}{an object of 'slidepzl_state' class.
It can have NAs in locations of pieces.}

\item{setting}{an object of 'slidepzl_setting' class.}

\item{verbose}{logical. Show messages?}
}
\value{
an integer.
        number of possible states.
}
\description{
return number of possible states.
It can take a long time.
}
