% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getShortestPaths}
\alias{getShortestPaths}
\title{get shortest paths}
\usage{
getShortestPaths(g, target = NULL)
}
\arguments{
\item{g}{an 'igraph' object generated by \code{makeGraph} function.}

\item{target}{a numerical vector, which gives
IDs of target states. NULL means IDs of goal states.}
}
\value{
a list with following elements:
        \describe{
          \item{state}{a list of character vectors.
          ith element gives states in ith pathes.}
          \item{transition}{a list of character vectors.
          ith element gives transitions in ith pathes.}
        }
}
\description{
Get shortest path from initial states to goal states
(or to specified states).
}
\details{
see \code{\link{makeGraph}} for example code.
}
