% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_state.R
\name{isGoalState}
\alias{isGoalState}
\title{check if a state satisfy goal conditions}
\usage{
isGoalState(state, goalcondition)
}
\arguments{
\item{state}{an object of 'slidepzl_state' class.}

\item{goalcondition}{an object of 'slidepzl_state' class.}
}
\value{
logical. Does the state include all pieces in goal conditions?
}
\description{
Internal. Check if a state satisfy goal conditions.
}
