% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isValidState}
\alias{isValidState}
\title{check the validity of state}
\usage{
isValidState(state, setting)
}
\arguments{
\item{state}{an object of 'slidepzl_state' class.}

\item{setting}{an object of 'slidepzl_setting' class.}
}
\value{
logical. is the state valid?
}
\description{
check the validity of state
}
