% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeNextStates}
\alias{makeNextStates}
\title{Internal. Return all states which can be generated by moving a piece of given state.}
\usage{
makeNextStates(state, setting)
}
\arguments{
\item{state}{an object of 'slidepzl_state' class. Original state.}

\item{setting}{an object of 'slidepzl_setting' class.}
}
\value{
a list of objects of 'slidepzl_state' class, giving
        all states which can be generated by moving a piece of
        the original state.
        Validity is guaranteed.

        The name of elements give how they are generated.
        e.g. '11D' means a piece at (1,1) in the original state is moved down.
}
\description{
Internal. Return all states which can be generated by moving a piece of given state.
}
