% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{makePiece}
\alias{makePiece}
\title{make a piece}
\usage{
makePiece(type, loc)
}
\arguments{
\item{type}{a alphabetic character giving 'type' of piece. Case sensitive.}

\item{loc}{a numeric vector of length 2,
giving the location (row, column) of the upper left corner
of the piece.
Should be >= 1 and <= 9, or NA.}
}
\value{
an object of class "slidepzl_piece", which is a list with following elements:
        \describe{
          \item{type}{'type' of piece}
          \item{loc}{location of piece}
        }
}
\description{
make a piece
}
\examples{
# a piece of type 'A' is on the board,
# whose upper left corner is located at (1, 1)
makePiece('A', c(1,1))
}
