% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeSetting}
\alias{makeSetting}
\title{make setting of a sliding puzzle}
\usage{
makeSetting(boardsize, piecesize)
}
\arguments{
\item{boardsize}{a numeric vector of length 2.
(number of rows, number of columns) of the board.
it should be >= 1 and <= 9.}

\item{piecesize}{a named list. it should have unduplicated names.
name:  a alphabetic character. case sensitive.
       'type' of piece.
value: a numeric vector of length 2.
       (vertical size, horizontal size) of the type of piece.}
}
\value{
an object of class "slidepzl_setting", which is a list with following elements:
        boardsize: (number of rows, number of columns) of the board.
        piecearea: a named list.
                   name:  'type' of piece.
                   value: a numeric matrix of size (k,2).
                          relative locations of cells which the type of piece occupies.
}
\description{
make setting of a sliding puzzle
}
\examples{
makeSPSetting(
  boardsize = c(3,3),
  piecesize = list(
    "A" = c(1,1),
    "B" = c(1,1),
    "C" = c(1,1),
    "D" = c(1,1),
    "E" = c(1,1),
    "F" = c(1,1),
    "G" = c(1,1)
  )
)
}
