% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeState}
\alias{makeState}
\title{make a state (or condition of states)}
\usage{
makeState(pieces)
}
\arguments{
\item{pieces}{a list of objects of 'slidepzl_piece' class.}
}
\value{
an object of class "slidepzl_state", which is a list with following elements:
        piecetypes: a character vector.
                    the ith element gives the 'type' of ith piece.
        piecelocs:  a numeric matrix of size (num.pieces, 2).
                    the ith row gives (row, column) of upper left corner of ith piece.
        validity is not guaranteed.
}
\description{
make a state (or condition of states)
}
