% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotGraph}
\alias{plotGraph}
\title{plot the network graph.}
\usage{
plotGraph(g, status_color = c("blue", "skyblue", "gray", "red"),
  status_label = FALSE, move_label = TRUE, method = "igraph", ...)
}
\arguments{
\item{g}{an object generated by \code{makeGraph} function.}

\item{status_color}{a vector of color, length 4.
Color of (initial, unexamined,
examined, goal) states (i.e. nodes).}

\item{status_label}{logical. show labels of states?}

\item{move_label}{logical. show labels of transition?}

\item{method}{'igraph': use igraph package; 'GGally': use GGally package.}

\item{...}{pass to \code{igraph::plot} or \code{GGally::ggnet2}}
}
\description{
Provide easy way to plot the network graph generated by
\code{makeGraph} function.
}
\details{
To plot with more detailed
specification, use \code{igraph::tkplot} function or
\code{GGally::ggnet2} function instead.

see \code{\link{makeGraph}} for example code.
}
