% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{as.character.slidepzl_piece}
\alias{as.character.slidepzl_piece}
\title{convert the location of a piece into a string}
\usage{
\method{as.character}{slidepzl_piece}(piece)
}
\arguments{
\item{piece}{a object of 'slidepzl_piece' class.}
}
\value{
a character string of length 3
}
\description{
Internal. Convert the location of a piece into a string.
}
