% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graph.R
\name{getShortestPath}
\alias{getShortestPath}
\title{get a shortest paths to target}
\usage{
getShortestPath(g, target = NULL)
}
\arguments{
\item{g}{an 'igraph' object generated by \code{makeGraph} function.}

\item{target}{a numerical vector, which gives
IDs of target states. NULL means IDs of goal states.}
}
\value{
a list with following elements:
        \describe{
          \item{state}{a character vector, giving states of the path.}
          \item{transition}{a character vectors,
          giving transitions of the path.}
        }
}
\description{
get one of the shortest path from initial states to goal states
(or to specified states).
}
