% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeNextStates}
\alias{makeNextStates}
\title{return possible next states from a state.}
\usage{
makeNextStates(state, setting)
}
\arguments{
\item{state}{an object of 'slidepzl_state' class.}

\item{setting}{an object of 'slidepzl_setting' class.}
}
\value{
a list of 'slidepzl_state' class.
        all of states which can be generated by moving a piece of
        the original state.
        their validity of guaranteed.
        the name of elements give how they are generated.
        e.g. '11D' means a piece at (1,1) in the original state is moved down.
}
\description{
return possible next states from a state.
}
