% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makePiece}
\alias{makePiece}
\title{make a piece}
\usage{
makePiece(type, loc)
}
\arguments{
\item{type}{a alphabetic character. case sensitive. giving 'type' of piece.}

\item{loc}{a numeric vector of length 2.
giving the location of the piece
((row, column) of the upper left corner of the piece).}
}
\value{
an object of class "slidepzl_piece", which is a list with following elements:
        type: 'type' of piece
        loc:  location of piece
}
\description{
make a piece
}
