% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeSetting}
\alias{makeSetting}
\title{make setting of a sliding puzzle}
\usage{
makeSetting(boardsize, piecesize)
}
\arguments{
\item{boardsize}{a numeric vector of length 2.
(number of rows, number of columns) of the board.
Should be >= 1 and <= 9.}

\item{piecesize}{a named list.
The name of each element should be a alphabetic character.
Case sensitive.
Duplicated names are not allowed.
Each element should be a numeric vector of length 2,
giving (vertical size, horizontal size) of the type of piece.}
}
\value{
an object of class 'slidepzl_setting',
        which is a list with following elements:
        \describe{
          \item{boardsize}{
            same as \code{boardsize} argument
          }
          \item{piecearea}{
            a named list.
            The name of each element is a alphabetic character,
            giving the 'type' of piece.
            Each element is a numeric matrix of size (k,2),
            giving relative locations of cells which that type of piece occupies.
          }
        }
}
\description{
make setting of a sliding puzzle
}
\examples{
mysetting <- makeSetting(
  boardsize = c(3,3),
  piecesize = list(
    "A" = c(1,1),
    "B" = c(1,1),
    "C" = c(1,1),
    "D" = c(1,1),
    "E" = c(1,1),
    "F" = c(1,1),
    "G" = c(1,1)
  )
)
}
