% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_state.R
\name{modifyState}
\alias{modifyState}
\title{generate new state by moving a piece in a state}
\usage{
modifyState(state, pieceid, diff)
}
\arguments{
\item{state}{an object of 'slidepzl_state' object.}

\item{pieceid}{integer. the piece which is moved.}

\item{diff}{a integer vector of length 2, giving the direction to move.}
}
\value{
an object of 'slidepzl_state' class, which can be invalid.
}
\description{
Internal. Generate new state by moving a piece in a state.
}
