% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modifyState}
\alias{modifyState}
\title{move a piece.}
\usage{
modifyState(state, pieceid, diff)
}
\arguments{
\item{state}{an object of 'slidepzl_state' object.}

\item{pieceid}{integer. the piece which is moved.}

\item{diff}{a integer vector of length 2. direction of moving.}
}
\value{
an object of 'slidepzl_state' class.
        its validity is not guaranteed.
}
\description{
move a piece.
}
