% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graph.R
\name{plotGraph}
\alias{plotGraph}
\title{plot a network graph}
\usage{
plotGraph(
  g,
  state_label = FALSE,
  state_status = TRUE,
  state_color = c("blue", "skyblue", "gray", "red"),
  move_label = TRUE,
  method = "igraph",
  community = NULL,
  weight_in_community = 5,
  fr_iter = 1000,
  ...
)
}
\arguments{
\item{g}{an object  generated by \code{makeGraph} function.}

\item{state_label}{Labels of states. It can be one of follows:
(1) a vector of which length is the number of vertexs in \code{g}.
(2) \code{TRUE}, which means \code{V(g)$name}.
(3) \code{FALSE}, which means no label.}

\item{state_status}{Status of states. It can be one of follows:
(1) a vector integer, of which length is the number of vertex in \code{g}.
(2) \code{TRUE}, which means \code{V(g)$status}.
(3) \code{FALSE}, which means no status.}

\item{state_color}{a vector of color names.
The ith state is colored by \code{state_color[state_status[i]]}.}

\item{move_label}{Labels of moves. It can be one of follows:
(1) a vector of which length is the number of edges in \code{g}.
(2) \code{TRUE}, which means \code{E(g)$name}.
(3) \code{FALSE}, which means no label.}

\item{method}{'igraph': use \code{igraph::plot}; 'GGally': use \code{GGally::ggnet2}}

\item{community}{an object of 'igraph::communities' class.
Community of \code{g}.
If specified, weights are assigned to edges
so that communities are separated each other.
It can be specified only when \code{method = 'igraph'}.}

\item{weight_in_community}{weights for edges within same community.
Bigger value makes communities to be separated much more.
Only effective when \code{community} is specified.}

\item{fr_iter}{iteration for fruchterman reingold algorithm.
Only effective when \code{community} is specified.}

\item{...}{pass to \code{igraph::plot} or \code{GGally::ggnet2}}
}
\description{
provide an easy way to plot a network graph returned by \code{makeGraph} function.
}
\details{
To plot with more detailed specification, use \code{igraph::tkplot} or
\code{GGally::ggnet2} function instead.
For huge graphs consider to use other software like Gephi.

see \code{\link{makeGraph}} for example code.
}
