% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_graph.R
\name{plotGraph}
\alias{plotGraph}
\title{plot a network graph}
\usage{
plotGraph(g, status_color = c("blue", "skyblue", "gray", "red"),
  status_label = FALSE, move_label = TRUE, method = "igraph", ...)
}
\arguments{
\item{g}{an object generated by \code{makeGraph} function.}

\item{status_color}{a vector of color, length 4.
Color of (initial, unexamined,
examined, goal) states}

\item{status_label}{logical. show labels of states?}

\item{move_label}{logical. show labels of transition?}

\item{method}{'igraph': use \code{igraph::plot}; 'GGally': use \code{GGally::ggnet2}}

\item{...}{pass to \code{igraph::plot} or \code{GGally::ggnet2}}
}
\description{
provide an easy way to plot a network graph returned by \code{makeGraph} function.
}
\details{
To plot with more detailed
specification, use \code{igraph::tkplot} or
\code{GGally::ggnet2} function instead.
For huge graphs consider to use other software like Gephi.

see \code{\link{makeGraph}} for example code.
}
